Refactor format util file
